refactor(default_chatter): 按组件和工具职责整理目录结构
