+++ plugins/default_chatter/components/__init__.py
@@ -1,0 +1,1 @@
+"""Default Chatter 组件实现。"""
