+++ plugins/default_chatter/components/config.py
