+++ plugins/default_chatter/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Default Chatter 可复用工具。"""
