+++ plugins/default_chatter/utils/multimodal.py
