+++ plugins/default_chatter/utils/prompts.py
